rename directories and files to normalize to module acronyms
